+++ src/lib/converter/comments/lineLexer.ts
@@ -169,7 +169,7 @@ function* lexLineComments2(
 
                 if (
                     lookahead !== pos + 1 &&
-                    (lookahead === end || /\s/.test(file[lookahead]))
+                    (lookahead === end || /[\s}]/.test(file[lookahead]))
                 ) {
                     braceStartsType = true;
                     yield makeToken(TokenSyntaxKind.Tag, lookahead - pos);