+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 -   Category children are now sorted according to the `sort` option, #2272.
 -   Inline tags no longer require a space after the tag name to be parsed as a tag, #2273.
+-   Fixed module/namespace links in navigation when viewed in Safari, #2275.
 
 ## v0.24.6 (2023-04-24)